TransactionService WIP.

+++ modules/dex-core/src/main/scala/org/ergoplatform/dex/domain/models/Order.scala
@@ -3,6 +3,13 @@ package org.ergoplatform.dex.domain.models
 import org.ergoplatform.dex.AssetId
 
 /** Global market order.
+  * @param `type` - type of the order (sell or buy)
+  * @param asset  - id of the asset X
+  * @param baseAsset - id of the asset Y
+  * @param amount - amount of the `asset`
+  * @param price - price for `asset` in `baseAsset`
+  * @param feePerToken - amount of fee (in nanoERG) per one traded `asset`
+  * @param meta - order metadata
   */
 final case class Order[T <: OrderType](
   `type`: T,
@@ -10,7 +17,7 @@ final case class Order[T <: OrderType](
   baseAsset: AssetId,
   amount: Long,
   price: Long,
-  feePerToken: Double,
+  feePerToken: Long,
   meta: OrderMeta
 )
 
@@ -25,7 +32,7 @@ object Order {
     baseAsset: AssetId,
     amount: Long,
     price: Long,
-    feePerToken: Double,
+    feePerToken: Long,
     meta: OrderMeta
   ): BuyOrder =
     Order(OrderType.Buy, asset, baseAsset, amount, price, feePerToken, meta)
@@ -35,7 +42,7 @@ object Order {
     baseAsset: AssetId,
     amount: Long,
     price: Long,
-    feePerToken: Double,
+    feePerToken: Long,
     meta: OrderMeta
   ): SellOrder =
     Order(OrderType.Sell, asset, baseAsset, amount, price, feePerToken, meta)
